requirements.txt
# ── Core runtime dependencies ─────────────────────────────────────────────── #
certifi==2026.2.25
charset-normalizer==3.4.7
idna==3.11
oauthlib==3.3.1
paho-mqtt==2.1.0
PyYAML>=6.0.2          # YAML rule file parsing (tesla/scheduler/loader.py)
requests==2.33.1
requests-oauthlib==2.0.0
TeslaPy==2.9.1
urllib3==2.6.3
websocket-client==1.9.0

# ── Development / testing ─────────────────────────────────────────────────── #
pytest>=8.0
